Rename erased task module
