+++ src/tasks/mod.rs
@@ -1,7 +1,7 @@
 pub(crate) mod context;
-pub(crate) mod erased;
+pub(crate) mod erased_task;
 pub mod task_trait;
 
 pub use context::Context;
-pub use erased::{ErasedTask, ErasedTaskFunction};
+pub use erased_task::{ErasedTask, ErasedTaskFunction};
 pub use task_trait::Task;
